+++ src/gt_logging.h
@@ -38,8 +38,8 @@
 #include "gt_logdestfunctor.h"
 
 // macros to convert an argument to string
-#define GT_LOG_TO_STR_HELPER(X) #X
-#define GT_LOG_TO_STR(X) GT_LOG_TO_STR_HELPER(X)
+#define GT_LOG_IMPL_TO_STR(X) #X
+#define GT_LOG_TO_STR(X) GT_LOG_IMPL_TO_STR(X)
 
 // logging id
 #ifndef GT_MODULE_ID
@@ -52,9 +52,43 @@ namespace gt
 namespace log
 {
 
+using hash_t = size_t;
+
+GT_LOGGING_EXPORT
+hash_t hash(std::string const& msg, std::string const& id, Level level);
+
+//! Main logger instance
 class Logger
 {
 public:
+
+    //! Default cache to store the hashes of the messages that should be
+    //! logged once. This cache is thread safe, thus may not have optimal
+    //! performance. Look up time is O(n).
+    class DefaultCache
+    {
+    public:
+        using value_type = hash_t;
+
+        GT_LOGGING_EXPORT DefaultCache();
+        GT_LOGGING_EXPORT ~DefaultCache();
+
+        //! Appends a new hash entry to the cache.
+        GT_LOGGING_EXPORT void append(hash_t hash);
+
+        //! Clears all saved hashes
+        GT_LOGGING_EXPORT void clear();
+
+        //! Returns whether the hash exists in the cache
+        GT_LOGGING_EXPORT bool find(hash_t hash) const noexcept;
+
+    private:
+        struct Impl;
+        std::unique_ptr<Impl> pimpl;
+    };
+
+    DefaultCache globalCache;
+
     GT_LOGGING_EXPORT
     static Logger& instance();
 
@@ -100,26 +134,29 @@ public:
     GT_LOGGING_EXPORT
     int verbosity() const;
 
+    //! Method to log a message to all destinations
+    GT_LOGGING_EXPORT
+    void log(Level level, std::string message, std::string id = GT_MODULE_ID);
+
     //! The helper forwards the streaming to QDebug and builds the final
     //! log message.
     class Helper
     {
     public:
 
-        explicit Helper(Level logLevel, std::string logId = GT_MODULE_ID) :
-            level{logLevel},
-            id{std::move(logId)}
+        explicit Helper(Level _level, std::string _id = GT_MODULE_ID) :
+            level{_level},
+            id{std::move(_id)}
         {}
 
         ~Helper() { writeToLog(); }
 
         gt::log::Stream& stream() { return gtStream; }
-
     private:
 
         Level level;
         std::string id;
-        gt::log::Stream gtStream;
+        Stream gtStream;
 
         GT_LOGGING_EXPORT void writeToLog();
     };
@@ -138,10 +175,60 @@ private:
     std::unique_ptr<Impl> pimpl;
 };
 
+template <typename Cache>
+class LogOnce
+{
+public:
+
+    explicit LogOnce(Cache& _cache, Level _level, std::string _id = GT_MODULE_ID) :
+        level{_level},
+        id{std::move(_id)},
+        cache(&_cache)
+    {}
+
+    LogOnce(LogOnce&&) = default;
+    ~LogOnce()
+    {
+        if (!gtStream.mayLog()) return;
+
+        std::string message = gtStream.str();
+        if (message.empty()) return;
+
+        hash_t hash = gt::log::hash(message, id, level);
+
+        bool hashExists = cache->find(hash);
+        if (hashExists) return;
+
+        cache->append(hash);
+        Logger::instance().log(level, std::move(message), std::move(id));
+    }
+
+    gt::log::Stream& stream() { return gtStream; }
+
+private:
+
+    Level level;
+    std::string id;
+    Stream gtStream;
+    Cache* cache;
+};
+
+//! Constructs a LogOnce Helper object that uses a custom cache
+template <typename Cache>
+inline auto logOnce(Cache& cache, Level level, std::string id  = GT_MODULE_ID)
+{
+    return LogOnce<Cache>(cache, level, std::move(id));
+}
+
+//! Constructs a LogOnce Helper object that uses the global cache
+inline auto logOnce(Level level, std::string id  = GT_MODULE_ID)
+{
+    return logOnce(Logger::instance().globalCache, level, std::move(id));
+}
+
 } // namespace log
 
 } // namespace gt
-
 
 // log line numbers
 #ifdef GT_LOG_LINE_NUMBERS
@@ -165,34 +252,86 @@ private:
 #define GT_LOG_IMPL_NOSPACE
 #endif
 
-#define GT_LOG_IMPL(LEVEL) \
-    if (gt::log::Logger::instance().loggingLevel() > gt::log::LEVEL) \
-    {} \
-    else gt::log::Logger::Helper(gt::log::LEVEL).stream() \
+////////// HELPER MACROS FOR COMON CODE //////////
+
+// log only if logging level matches
+#define GT_LOG_IMPL_IF_LEVEL(LEVEL) \
+    if (gt::log::Logger::instance().loggingLevel() <= gt::log::LEVEL)
+
+// apply global flags (quote, nospace, line numbers)
+#define GT_LOG_IMPL_APPLY_FLAGS() \
     GT_LOG_IMPL_LINE_NUMBERS GT_LOG_IMPL_QUOTE GT_LOG_IMPL_NOSPACE
 
-#define GT_LOG_IMPL_ID(LEVEL, ID) \
-    if (gt::log::Logger::instance().loggingLevel() > gt::log::LEVEL) \
-    {} \
-    else gt::log::Logger::Helper(gt::log::LEVEL, ID).stream() \
-    GT_LOG_IMPL_LINE_NUMBERS GT_LOG_IMPL_QUOTE GT_LOG_IMPL_NOSPACE
+#define GT_LOG_IMPL_NO_ARG_EXPANDER() ,,
 
+////////// DEFAULT LOGGING MACROS //////////
 
-//! Default logging macros.
-#define gtTrace()       GT_LOG_IMPL(TraceLevel)
-#define gtDebug()       GT_LOG_IMPL(DebugLevel)
-#define gtInfo()        GT_LOG_IMPL(InfoLevel)
-#define gtWarning()     GT_LOG_IMPL(WarnLevel)
-#define gtError()       GT_LOG_IMPL(ErrorLevel)
-#define gtFatal()       GT_LOG_IMPL(FatalLevel)
+#define GT_LOG_IMPL_MESSAGE(LEVEL) \
+    GT_LOG_IMPL_IF_LEVEL(LEVEL) \
+        gt::log::Logger::Helper(gt::log::LEVEL).stream() \
+            GT_LOG_IMPL_APPLY_FLAGS()
 
-//! Logging macros with custom logging id
-#define gtTraceId(ID)   GT_LOG_IMPL_ID(TraceLevel, ID)
-#define gtDebugId(ID)   GT_LOG_IMPL_ID(DebugLevel, ID)
-#define gtInfoId(ID)    GT_LOG_IMPL_ID(InfoLevel , ID)
-#define gtWarningId(ID) GT_LOG_IMPL_ID(WarnLevel , ID)
-#define gtErrorId(ID)   GT_LOG_IMPL_ID(ErrorLevel, ID)
-#define gtFatalId(ID)   GT_LOG_IMPL_ID(FatalLevel, ID)
+//! Default logging macros
+#define gtTrace()       GT_LOG_IMPL_MESSAGE(TraceLevel)
+#define gtDebug()       GT_LOG_IMPL_MESSAGE(DebugLevel)
+#define gtInfo()        GT_LOG_IMPL_MESSAGE(InfoLevel)
+#define gtWarning()     GT_LOG_IMPL_MESSAGE(WarningLevel)
+#define gtError()       GT_LOG_IMPL_MESSAGE(ErrorLevel)
+#define gtFatal()       GT_LOG_IMPL_MESSAGE(FatalLevel)
+
+////////// DEFAULT LOGGING MACROS WITH ID //////////
+
+#define GT_LOG_IMPL_MEESAGE_ID(LEVEL, ID) \
+    GT_LOG_IMPL_IF_LEVEL(LEVEL) \
+        gt::log::Logger::Helper(gt::log::LEVEL, ID).stream() \
+            GT_LOG_IMPL_APPLY_FLAGS()
+
+#define gtTraceId(ID)   GT_LOG_IMPL_MEESAGE_ID(TraceLevel, ID)
+#define gtDebugId(ID)   GT_LOG_IMPL_MEESAGE_ID(DebugLevel, ID)
+#define gtInfoId(ID)    GT_LOG_IMPL_MEESAGE_ID(InfoLevel , ID)
+#define gtWarningId(ID) GT_LOG_IMPL_MEESAGE_ID(WarningLevel , ID)
+#define gtErrorId(ID)   GT_LOG_IMPL_MEESAGE_ID(ErrorLevel, ID)
+#define gtFatalId(ID)   GT_LOG_IMPL_MEESAGE_ID(FatalLevel, ID)
+
+////////// LOG ONCE MACRO //////////
+
+#define GT_LOG_IMPL_ONCE_F1(LEVEL, ...) \
+    GT_LOG_IMPL_IF_LEVEL(LEVEL ## Level) \
+        gt::log::logOnce(gt::log::LEVEL ## Level).stream() \
+            GT_LOG_IMPL_APPLY_FLAGS()
+#define GT_LOG_IMPL_ONCE_F2(LEVEL, CACHE, ...) \
+    GT_LOG_IMPL_IF_LEVEL(LEVEL ## Level) \
+        gt::log::logOnce(CACHE, gt::log::LEVEL ## Level).stream() \
+            GT_LOG_IMPL_APPLY_FLAGS()
+
+// variadic macro magic
+#define GT_LOG_IMPL_ONCE_FUNC_CHOOSER(_f1, _f2, _f3, ...) _f3
+#define GT_LOG_IMPL_ONCE_FUNC_RECOMPOSER(ArgsWithBrackets) GT_LOG_IMPL_ONCE_FUNC_CHOOSER ArgsWithBrackets
+#define GT_LOG_IMPL_ONCE_CHOOSE_FROM_ARG_COUNT(...) GT_LOG_IMPL_ONCE_FUNC_RECOMPOSER((__VA_ARGS__, GT_LOG_IMPL_ONCE_F2, GT_LOG_IMPL_ONCE_F1, ))
+#define GT_LOG_IMPL_ONCE_MACRO_CHOOSER(...) GT_LOG_IMPL_ONCE_CHOOSE_FROM_ARG_COUNT(GT_LOG_IMPL_NO_ARG_EXPANDER __VA_ARGS__ ())
+
+#define gtLogOnce(...) GT_LOG_IMPL_ONCE_MACRO_CHOOSER( __VA_ARGS__)( __VA_ARGS__)
+
+////////// LOG ONCE WITH ID MACRO //////////
+
+#define GT_LOG_IMPL_ONCE_ID_F2(LEVEL, ID, ...) \
+    GT_LOG_IMPL_IF_LEVEL(LEVEL ## Level) \
+        gt::log::logOnce(gt::log::LEVEL ## Level, ID).stream() \
+            GT_LOG_IMPL_APPLY_FLAGS()
+#define GT_LOG_IMPL_ONCE_ID_F3(LEVEL, ID, CACHE, ...) \
+    GT_LOG_IMPL_IF_LEVEL(LEVEL ## Level) \
+        gt::log::logOnce(CACHE, gt::log::LEVEL ## Level, ID).stream() \
+            GT_LOG_IMPL_APPLY_FLAGS()
+
+// variadic macro magic
+#define GT_LOG_IMPL_ONCE_ID_FUNC_CHOOSER(_f1, _f2, _f3, _f4, ...) _f4
+#define GT_LOG_IMPL_ONCE_ID_FUNC_RECOMPOSER(ArgsWithBrackets) GT_LOG_IMPL_ONCE_ID_FUNC_CHOOSER ArgsWithBrackets
+#define GT_LOG_IMPL_ONCE_ID_CHOOSE_FROM_ARG_COUNT(...) GT_LOG_IMPL_ONCE_ID_FUNC_RECOMPOSER((__VA_ARGS__, GT_LOG_IMPL_ONCE_ID_F3, GT_LOG_IMPL_ONCE_ID_F2, ,))
+#define GT_LOG_IMPL_ONCE_ID_MACRO_CHOOSER(...) GT_LOG_IMPL_ONCE_ID_CHOOSE_FROM_ARG_COUNT(GT_LOG_IMPL_NO_ARG_EXPANDER __VA_ARGS__ ())
+
+#define gtLogOnceId(...) GT_LOG_IMPL_ONCE_ID_MACRO_CHOOSER( __VA_ARGS__)( __VA_ARGS__)
+
+////////// APPLY GLOBAL DEFINES //////////
 
 #ifdef GT_LOG_DISABLE
 #include "gt_logdisablelogforfile.h"